+++ game_patch/misc/alpine_settings.h
@@ -126,6 +126,7 @@ struct AlpineGameSettings
     bool entity_pain_sounds = true;
     bool real_armor_values = false;
     bool always_show_spectators = false;
+    bool simple_server_chat_msgs = true;
     bool quick_exit = false;
     int colorblind_mode = 0; // 0=off,1=protanopia,2=deuteranopia,3=tritanopia
     int suppress_autoswitch_alias = -1;