+++ game_patch/multi/alpine_packets.cpp
@@ -17,6 +17,8 @@
 #include "alpine_packets.h"
 #include "../misc/player.h"
 #include "../hud/hud.h"
+#include "../sound/sound.h"
+#include "../misc/alpine_settings.h"
 
 void af_send_packet(rf::Player* player, const void* data, int len, bool is_reliable)
 {
@@ -1417,28 +1419,28 @@ void af_send_server_cfg(rf::Player* player) {
     std::string output{};
     print_alpine_dedicated_server_config_info(output, true);
 
-    const auto send_msg = [player] (const std::string_view data) {
+    const auto send_msg = [player] (const std::string_view msg) {
         constexpr size_t max_chunk_len = rf::max_packet_size - sizeof(af_server_msg_packet);
-        const size_t len = std::clamp(data.size(), 0uz, max_chunk_len);
+        const size_t len = std::clamp(msg.size(), 0uz, max_chunk_len);
 
-        af_server_msg_packet msg_packet;
-        msg_packet.header.type = static_cast<uint8_t>(af_packet_type::af_server_msg);
-        msg_packet.header.size = static_cast<uint16_t>(
-            sizeof(msg_packet)
-                - sizeof(msg_packet.header)
+        af_server_msg_packet server_msg_packet;
+        server_msg_packet.header.type = static_cast<uint8_t>(af_packet_type::af_server_msg);
+        server_msg_packet.header.size = static_cast<uint16_t>(
+            sizeof(server_msg_packet)
+                - sizeof(server_msg_packet.header)
                 + len
         );
-        msg_packet.type = static_cast<uint8_t>(AF_SERVER_MSG_TYPE_REMOTE_SERVER_CFG);
+        server_msg_packet.type = static_cast<uint8_t>(AF_SERVER_MSG_TYPE_REMOTE_SERVER_CFG);
 
-        std::array<uint8_t, sizeof(msg_packet) + max_chunk_len> buf{}; 
-        std::memcpy(buf.data(), &msg_packet, sizeof(msg_packet));
-        uint8_t* msg = buf.data() + sizeof(msg_packet);
-        std::memcpy(msg, data.data(), len);
+        std::array<uint8_t, sizeof(server_msg_packet) + max_chunk_len> buf{}; 
+        std::memcpy(buf.data(), &server_msg_packet, sizeof(server_msg_packet));
+        uint8_t* ptr = buf.data() + sizeof(server_msg_packet);
+        std::memcpy(ptr, msg.data(), len);
 
         rf::multi_io_send_reliable(
             player,
             buf.data(),
-            msg_packet.header.size + sizeof(msg_packet.header),
+            server_msg_packet.header.size + sizeof(server_msg_packet.header),
             0
         );
 
@@ -1448,6 +1450,76 @@ void af_send_server_cfg(rf::Player* player) {
     constexpr int chunk_size = rf::max_packet_size - sizeof(af_server_msg_packet);
     for (const auto chunk : output | std::views::chunk(chunk_size)) {
         send_msg(std::string_view{chunk.begin(), chunk.end()});
+    }
+}
+
+union af_server_msg_packet_buf {
+    af_server_msg_packet packet;
+    std::array<uint8_t, rf::max_packet_size> buf;
+};
+
+af_server_msg_packet_buf build_automated_chat_msg_packet(
+    const std::string_view msg
+) {
+    constexpr size_t max_len = rf::max_packet_size - sizeof(af_server_msg_packet);
+    const size_t len = std::clamp(msg.size(), 0uz, max_len);
+
+    af_server_msg_packet_buf buf{};
+    buf.packet.header.type = static_cast<uint8_t>(af_packet_type::af_server_msg);
+    buf.packet.header.size = static_cast<uint16_t>(
+        sizeof(buf.packet)
+            - sizeof(buf.packet.header)
+            + len
+    );
+    buf.packet.type = static_cast<uint8_t>(AF_SERVER_MSG_TYPE_AUTOMATED_CHAT);
+    std::memcpy(buf.packet.data, msg.data(), len);
+
+    return buf;
+}
+
+void af_broadcast_automated_chat_msg(const std::string_view msg) {
+    if (!rf::is_multi || !rf::is_server) {
+        return;
+    }
+
+    rf::console::print("Server: {}", msg);
+
+    const af_server_msg_packet_buf buf = build_automated_chat_msg_packet(msg);
+
+    for (rf::Player& player : SinglyLinkedList{rf::player_list}) {
+        if (&player == rf::local_player) {
+            continue;
+        }
+
+        if (is_player_minimum_af_client_version(&player, 1, 2)) {
+            rf::multi_io_send_reliable(
+                &player,
+                &buf.packet,
+                buf.packet.header.size + sizeof(buf.packet.header),
+                0
+            );
+        } else {
+            send_chat_line_packet(msg, &player);
+        }
+    }
+}
+
+void af_send_automated_chat_msg(const std::string_view msg, rf::Player* player) {
+    if (!rf::is_multi || !rf::is_server || !player) {
+        return;
+    }
+
+    if (is_player_minimum_af_client_version(player, 1, 2)) {
+        const af_server_msg_packet_buf buf = build_automated_chat_msg_packet(msg);
+
+        rf::multi_io_send_reliable(
+            player,
+            &buf.packet,
+            buf.packet.header.size + sizeof(buf.packet.header),
+            0
+        );
+    } else {
+        send_chat_line_packet(msg, player);
     }
 }
 
@@ -1469,10 +1541,17 @@ void af_process_server_msg_packet(
     std::memcpy(&msg_packet, data, sizeof(msg_packet));
 
     if (msg_packet.type == static_cast<uint8_t>(AF_SERVER_MSG_TYPE_REMOTE_SERVER_CFG)) {
-        const char* msg = static_cast<const char*>(data) + sizeof(msg_packet);
+        const char* ptr = static_cast<const char*>(data) + sizeof(msg_packet);
         g_remote_server_cfg_popup.add_content(
-            std::string_view{msg, len - sizeof(msg_packet)}
+            std::string_view{ptr, len - sizeof(msg_packet)}
         );
+    } else if (msg_packet.type == static_cast<uint8_t>(AF_SERVER_MSG_TYPE_AUTOMATED_CHAT)) {
+        const char* ptr = static_cast<const char*>(data) + sizeof(msg_packet);
+        const rf::String msg{std::string_view{ptr, len - sizeof(msg_packet)}};
+        rf::multi_chat_print(msg, rf::ChatMsgColor::gold_white, rf::String{"Server: "});
+        if (!g_alpine_game_config.simple_server_chat_msgs) {
+            rf::snd_play(4, 0, 0.f, 1.f);
+        }
     }
 }