+++ game_patch/multi/alpine_packets.h
@@ -196,6 +196,7 @@ struct af_spectate_notify_packet {
 
 enum af_server_msg_type : uint8_t {
     AF_SERVER_MSG_TYPE_REMOTE_SERVER_CFG = 0x1,
+    AF_SERVER_MSG_TYPE_AUTOMATED_CHAT = 0x2,
 };
 
 struct af_server_msg_packet {
@@ -238,6 +239,8 @@ void af_send_spectate_notify_packet(rf::Player* player, const rf::Player* specta
 void af_process_spectate_notify_packet(const void* data, size_t len, const rf::NetAddr&);
 void af_send_server_cfg(rf::Player* player);
 void af_process_server_msg_packet(const void* data, size_t len, const rf::NetAddr&);
+void af_broadcast_automated_chat_msg(std::string_view msg);
+void af_send_automated_chat_msg(std::string_view msg, rf::Player* player);
 
 // client requests
 void af_send_handicap_request(uint8_t amount);